Move global types inside packages
