+++ packages/edtr-services/src/context/test/useDomTestData.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { eventEditor, currentUser, generalSettings } from './data';
 
 import { mockEspressoDomData } from '@eventespresso/services/src/config/test/data';
-import { EventEspressoData } from '../../../../../types';
+import { EventEspressoData } from '../../../../types';
 
 const useDomTestData = (): void => {
 	// Set the DOM data
