+++ packages/edtr-services/src/index.ts
@@ -1,5 +1,5 @@
 // import global types.
-import '../../../types';
+import '../../types';
 
 export * from './apollo';
