+++ packages/i18n/src/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @wordpress/i18n-no-variables */
 /* eslint-disable @wordpress/i18n-text-domain */
 // import global types.
-import '../../../types';
+import '../../types';
 import { createI18n } from '@wordpress/i18n';
 
 const i18nData = window?.eventEspressoData?.i18n;
